LayoutTests/storage/indexeddb/migrate-basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test migration from SQLite to LevelDB.");

if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    if (window.layoutTestController)
        layoutTestController.setOverrideIndexedDBBackingStore("sqlite");

    request = evalAndLog("webkitIndexedDB.open('MigrationDatabase')");
    request.onsuccess = sqliteOpenSuccess;
    request.onerror = unexpectedErrorCallback;
}

function sqliteOpenSuccess()
{
    debug("sqliteCreateSuccess():");
    window.db = evalAndLog("db = event.target.result");

    request = evalAndLog("db.setVersion('new version')");
    request.onsuccess = sqliteSetVersionSuccess;
    request.onerror = unexpectedErrorCallback;
}

function sqliteSetVersionSuccess()
{
    debug("sqliteSetVersionSuccess():");
    evalAndLog("window.db.createObjectStore('PlainOldStore', {autoIncrement: false})");
    window.store = evalAndLog("store = window.db.createObjectStore('StoreWithKeyPath', {keyPath: 'id', autoIncrement: true})");
    evalAndLog("window.db.createObjectStore('StoreWithAutoIncrement', {autoIncrement: true})");

    window.trans = evalAndLog("trans = db.transaction([], webkitIDBTransaction.READ_WRITE)");
    window.index = evalAndLog("index = store.createIndex('ExampleIndex','id', false)");


    sqliteTestAddRecords1();
}

function sqliteTestAddRecords1()
{
    debug("sqliteTestAddRecords1():");
    window.store = evalAndLog("store = trans.objectStore('PlainOldStore')");
    request = evalAndLog("store = store.add({name: 'George'},1)");
    request.onsuccess = sqliteTestAddRecords2;
    request.onfailure = unexpectedErrorCallback;
}

function sqliteTestAddRecords2()
{
    debug("sqliteTestAddRecords2():");
    window.store = evalAndLog("store = trans.objectStore('StoreWithKeyPath')");

    request = evalAndLog("store.add({name: 'Thomas', id: 3})");
    request.onsuccess = sqliteTestAddRecords3;
    request.onfailure = unexpectedErrorCallback;
}

function sqliteTestAddRecords3()
{
    debug("sqliteTestAddRecords3():");
    window.store = evalAndLog("store = trans.objectStore('StoreWithAutoIncrement')");
    request = evalAndLog("store.add({name: 'Lincoln', number: '7012'}, 5)");
    request.onsuccess = sqliteCloseDatabase;
    request.onfailure = unexpectedErrorCallback;
}


function sqliteCloseDatabase()
{
    window.db.close();
    window.db = null;
    window.trans = null;
    leveldbOpen();
}

function leveldbOpen()
{
    if (window.layoutTestController)
        layoutTestController.setOverrideIndexedDBBackingStore("leveldb");

    // This next line will trigger the migration.
    request = webkitIndexedDB.open('MigrationDatabase');
    request.onsuccess = leveldbOpenSuccess;
    request.onerror = unexpectedErrorCallback;
}

function leveldbOpenSuccess()
{
    debug("leveldbOpenSuccess():");
    window.db = evalAndLog("db = event.target.result");
    window.trans = evalAndLog("trans = db.transaction([], webkitIDBTransaction.READ_WRITE, 0)");
    leveldbCheckPlainOldStore();
}

function leveldbCheckPlainOldStore()
{
    debug("leveldbCheckPlainOldStore():");
    window.store = evalAndLog("store = trans.objectStore('PlainOldStore')");
    window.keyRange = webkitIDBKeyRange.lowerBound(0);
    request = evalAndLog("request = store.openCursor(keyRange)");
    if (!request) {
        testMigrationEnd();
        return;
    }
    request.onsuccess = leveldbCheckPlainOldStoreCursorNext;
    request.onerror = unexpectedErrorCallback;
}

function leveldbCheckPlainOldStoreCursorNext()
{
    debug("leveldbCheckPlainOldStoreCursorNext():");
    window.cursor = event.target.result;
    if (!window.cursor) {
        leveldbCheckStoreWithKeyPath();
        return;
    }

    shouldBeTrue("cursor.value.name == 'George'");
    cursor.continue();
}

function leveldbCheckStoreWithKeyPath()
{
    debug("leveldbCheckStoreWithKeyPath():");
    window.store = evalAndLog("store = trans.objectStore('StoreWithKeyPath')");
    var keyRange = webkitIDBKeyRange.lowerBound(0);
    request = store.openCursor(keyRange);
    request.onsuccess = leveldbCheckStoreWithKeyPathCursorNext;
    request.onerror = unexpectedErrorCallback;
}

function leveldbCheckStoreWithKeyPathCursorNext()
{
    debug("leveldbCheckStoreWithKeyPathCursorNext():");
    window.cursor = event.target.result;
    if (!window.cursor) {
        leveldbCheckStoreWithAutoIncrement();
        return;
    }

    shouldBeTrue("cursor.value.name == 'Thomas'");
    shouldBeTrue("cursor.value.id == '3'");
    cursor.continue();
}

function leveldbCheckStoreWithAutoIncrement()
{
    debug("leveldbCheckStoreWithAutoIncrement():");
    window.store = evalAndLog("store = trans.objectStore('StoreWithAutoIncrement')");
    var keyRange = webkitIDBKeyRange.lowerBound(0);
    request = store.openCursor(keyRange);
    request.onsuccess = leveldbCheckStoreWithAutoIncrementCursorNext;
    request.onerror = unexpectedErrorCallback;
}
function leveldbCheckStoreWithAutoIncrementCursorNext()
{
    debug("leveldbCheckStoreWithAutoIncrementCursorNext():");
    window.cursor = event.target.result;
    if (!window.cursor) {
        testMigrationEnd();
        return;
    }

    shouldBeTrue("cursor.value.name == 'Lincoln'");
    shouldBeTrue("cursor.value.number == '7012'");
    cursor.continue();
}

function testMigrationEnd()
{
    window.db.close();
    window.db = null;
    layoutTestController.clearAllDatabases();
    layoutTestController.setOverrideIndexedDBBackingStore("default");
    done();
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
