LayoutTests/storage/indexeddb/migrate-basics-expected.txt

Test migration from SQLite to LevelDB.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


webkitIndexedDB.open('MigrationDatabase')
sqliteCreateSuccess():
db = event.target.result
db.setVersion('new version')
sqliteSetVersionSuccess():
window.db.createObjectStore('MigrationPlainOldStore', {autoIncrement: false})
store = window.db.createObjectStore('MigrationStoreWithKeyPath', {keyPath: 'id', autoIncrement: true})
window.db.createObjectStore('MigrationStoreWithAutoIncrement', {autoIncrement: true})
trans = db.transaction([], webkitIDBTransaction.READ_WRITE)
index = store.createIndex('ExampleIndex','id', false)
sqliteTestAddRecords1():
store = trans.objectStore('MigrationPlainOldStore')
store = store.add({name: 'George'},1)
sqliteTestAddRecords2():
store = trans.objectStore('MigrationStoreWithKeyPath')
store.add({name: 'Thomas', id: 3})
sqliteTestAddRecords3():
store = trans.objectStore('MigrationStoreWithAutoIncrement')
store.add({name: 'Lincoln', number: '7012'}, 5)
leveldbOpenSuccess():
db = event.target.result
trans = db.transaction([], webkitIDBTransaction.READ_WRITE, 0)
leveldbCheckPlainOldStore():
store = trans.objectStore('MigrationPlainOldStore')
request = store.openCursor(keyRange)
leveldbCheckPlainOldStoreCursorNext():
PASS cursor.value.name == 'George' is true
leveldbCheckPlainOldStoreCursorNext():
leveldbCheckStoreWithKeyPath():
store = trans.objectStore('MigrationStoreWithKeyPath')
leveldbCheckStoreWithKeyPathCursorNext():
PASS ((window.index = window.store.index('ExampleIndex')) != undefined) is true
PASS (window.keyIndexCursor.key == 3) is true
PASS cursor.value.name == 'Thomas' is true
PASS cursor.value.id == '3' is true
leveldbCheckStoreWithKeyPathCursorNext():
leveldbCheckStoreWithAutoIncrement():
store = trans.objectStore('MigrationStoreWithAutoIncrement')
leveldbCheckStoreWithAutoIncrementCursorNext():
PASS cursor.value.name == 'Lincoln' is true
PASS cursor.value.number == '7012' is true
leveldbCheckStoreWithAutoIncrementCursorNext():
PASS successfullyParsed is true

TEST COMPLETE
